requirements.txt
bcrypt==3.2.0
click==8.1.2
python-dotenv==0.20.0
starlette==0.20.0
uvicorn==0.17.6
pyjwt==2.4.0
starlette-jwt==0.1.9
pydantic~=1.9.1